django_socketio/requirements.txt
Django==4.0.5
python-engineio==4.3.2
python-socketio==5.6.0
django-cors-headers==3.13.0